⚡ Optimize transliterate and sanitize_name allocations
- Refactor `transliterate` and `sanitize_name` to use `transliterate_into` helper
- Pre-allocate String buffers with capacity to avoid reallocations
- Avoid `format!` macro in `sanitize_name`
- Fix regression test visibility by moving `tests/regression_operator_drop.rs` to `src/semantic/expressions.rs`
